Rename internal package name to internals
