+++ balloon/src/main/kotlin/com/skydoves/balloon/internals/ActivityBalloonLazy.kt
@@ -14,24 +14,27 @@
  * limitations under the License.
  */
 
-package com.skydoves.balloon.internal
+package com.skydoves.balloon.internals
 
-import androidx.fragment.app.Fragment
+import android.content.Context
+import androidx.lifecycle.LifecycleOwner
 import com.skydoves.balloon.Balloon
 import java.io.Serializable
 import kotlin.reflect.KClass
 
 /**
- * An implementation of [Lazy] for creating an instance of the [Balloon] in Fragments.
- * Tied to the given fragment's lifecycle and, [factory].
+ * An implementation of [Lazy] for creating an instance of the [Balloon] lazily in Activities.
+ * Tied to the given [lifecycleOwner], [factory].
  *
- * @param fragment An instance of the [Balloon] will be created in this Fragment lazily.
+ * @param context A context for creating resources of the [Balloon] lazily.
+ * @param lifecycleOwner A [LifecycleOwner] for dismissing automatically when the [LifecycleOwner] is being destroyed.
  * This will prevents memory leak: [Avoid Memory Leak](https://github.com/skydoves/balloon#avoid-memory-leak).
  * @param factory A [Balloon.Factory] kotlin class for creating a new instance of the Balloon.
  */
 @PublishedApi
-internal class FragmentBalloonLazy<out T : Balloon.Factory>(
-  private val fragment: Fragment,
+internal class ActivityBalloonLazy<out T : Balloon.Factory>(
+  private val context: Context,
+  private val lifecycleOwner: LifecycleOwner,
   private val factory: KClass<T>
 ) : Lazy<Balloon>, Serializable {
 
@@ -41,26 +44,15 @@ internal class FragmentBalloonLazy<out T : Balloon.Factory>(
     get() {
       var instance = cached
       if (instance === null) {
-        if (fragment.context != null) {
-          val factory = factory::java.get().newInstance()
-          val lifecycle = if (fragment.view !== null) {
-            fragment.viewLifecycleOwner
-          } else {
-            fragment
-          }
-          instance = factory.create(fragment.requireActivity(), lifecycle)
-          cached = instance
-        } else {
-          throw IllegalArgumentException(
-            "Balloon can not be initialized. The passed fragment's context is null."
-          )
-        }
+        val factory = factory::java.get().newInstance()
+        instance = factory.create(context, lifecycleOwner)
+        cached = instance
       }
 
       return instance
     }
 
-  override fun isInitialized() = cached !== null
+  override fun isInitialized(): Boolean = cached !== null
 
   override fun toString(): String =
     if (isInitialized()) value.toString() else "Lazy value not initialized yet."
