+++ balloon/src/main/kotlin/com/skydoves/balloon/vectortext/VectorTextView.kt
@@ -21,8 +21,8 @@ import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
 import com.skydoves.balloon.R
 import com.skydoves.balloon.extensions.applyDrawable
-import com.skydoves.balloon.internal.NO_INT_VALUE
-import com.skydoves.balloon.internal.takeIfNotNoIntValue
+import com.skydoves.balloon.internals.NO_INT_VALUE
+import com.skydoves.balloon.internals.takeIfNotNoIntValue
 
 /** VectorTextView is a customizable textView having a vector icon.  */
 public class VectorTextView @JvmOverloads constructor(
